(BSR)[API] feat: move coll. functions from offers.api
Some collective offers related functions were inside the offers.api
module. It was not logical.

Now they play nicely with their friends from the educational.api.offer
module.
